cloud_function/requirements.txt
flask
google-cloud-aiplatform[pipelines]
